requirements.txt
PyQt6==6.6.1
PyInstaller==6.13.0
PyYAML==6.0.2
pyyaml==6.0.3
pycryptodomex==3.23.0